backend/requirements.txt
flask>=3.0.0
flask-socketio>=5.3.6
flask-cors>=4.0.0
anthropic>=0.52.0
python-dotenv>=1.0.0
supabase>=2.3.0
beautifulsoup4>=4.12.3
selenium>=4.18.1
webdriver-manager>=4.0.1
requests>=2.31.0
python-jose>=3.3.0
simple-websocket>=1.0.0
httpx>=0.27.0
duckduckgo-search>=7.0.0
pypdf>=4.2.0
openai>=1.30.0
# TEEPO Exam additions
pdfplumber>=0.11.0
pytesseract>=0.3.13
pillow>=10.0.0
